+++ CHANGELOG.md
@@ -1,5 +1,17 @@
 # Changelog
 
+## [1.2.0]
+
+### Added
+
+- ```Get-OPSSLProducts```
+  - Get a list of all the SSL products
+- ```Get-OPSSLOrders```
+  - Get a list of all orders or use the ```-ExpiringSoon``` switch to list the SSL orders that are expiring soon (30 days)
+- ```Send-OPSSLApproverEmail```
+  - (Re)send approve email to the contact, output will specify which email address this is.
+- Verbose ouputs for the request splat in ```Invoke-OPRequest```
+
 ## [v1.1.1]
 
 ### Added
@@ -24,7 +36,7 @@
 - Added ```$true``` outputs for all Get, Set and Remove commands when completed succesfully for logic.
 - ```[CmdletBinding()]``` to every function for common parameters.
 - ```Connect-OpenProvider``` now uses ```Invoke-OPRequest```
-- More readable and useful outputs for all commands that return information (```Get-OP*```)
+- More readable and useful outputs for all commands that return information (```Get-OP-```)
 - ```Update-OPDomain``` add comments, lock the domain and change nameserver groups with this command.
 - ```Get-OPNameserverGroup``` get a list of nameserver groups in your tenant.
 - ```New-OPAuthCode``` get a AuthCode for your domain (specify domain ID)